Update README and SYSTEM_STATUS; optimize CacheWarmingService and InMemoryEventBus
- Updated README.md to reflect changes in technology stack (Java 21, ClickHouse) and local setup instructions.
- Added SYSTEM_STATUS.md to provide a comprehensive report on the current system setup and status.
- Optimized InMemoryEventBus for type safety with generics.
- Enhanced CacheWarmingService for better performance under high-frequency trading (HFT) conditions, including pre-sizing maps and reducing lambda allocations.
- Improved unit tests for CacheWarmingService to ensure low latency impact and validate performance targets.
- Refactored PostgresOrderStateStore for cleaner SQL statements and improved error handling.
- Introduced application.properties for configuration management of ClickHouse and Kafka.
- Minor adjustments in MappedOrderStateStore for clarity and performance.

+++ core/orderbook/src/main/java/com/thelastwar/orderbook/MappedOrderStateStore.java
@@ -193,7 +193,7 @@ public class MappedOrderStateStore implements OrderStateStore {
                 mappedBuffer.putInt(0); // Write count = 0
                 // Zero out the data area to prevent stale data
                 byte[] zeros = new byte[1024];
-                int remaining = (int) Math.min(mappedBuffer.remaining(), ORDER_SIZE * MAX_ORDERS);
+                int remaining = Math.min(mappedBuffer.remaining(), ORDER_SIZE * MAX_ORDERS);
                 while (remaining > 0) {
                     int toWrite = Math.min(remaining, zeros.length);
                     mappedBuffer.put(zeros, 0, toWrite);
